move asn1 private key processing into separate module
